+++ tests/unit/lucatume/WPBrowser/Events/DispatcherTest.php
@@ -5,7 +5,7 @@ namespace lucatume\WPBrowser\Events;
 
 use Codeception\Events;
 use Codeception\Test\Unit;
-use lucatume\WPBrowser\Tests\Traits\UopzFunctions;
+use lucatume\WPBrowser\Traits\UopzFunctions;
 use Psr\EventDispatcher\EventDispatcherInterface;
 use Symfony\Component\EventDispatcher\EventDispatcher;