+++ tests/unit/lucatume/WPBrowser/WordPress/InstallationState/EmptyDirTest.php
@@ -5,13 +5,16 @@ namespace lucatume\WPBrowser\WordPress\InstallationState;
 
 use Codeception\Test\Unit;
 use lucatume\WPBrowser\Tests\Traits\TmpFilesCleanup;
-use lucatume\WPBrowser\Tests\Traits\UopzFunctions;
+use lucatume\WPBrowser\Traits\UopzFunctions;
 use lucatume\WPBrowser\Utils\Env;
 use lucatume\WPBrowser\Utils\Filesystem as FS;
 use lucatume\WPBrowser\Utils\Random;
 use lucatume\WPBrowser\WordPress\Database\MysqlDatabase;
 use lucatume\WPBrowser\WordPress\InstallationException;
 
+/**
+ * @group slow
+ */
 class EmptyDirTest extends Unit
 {
     use UopzFunctions;
@@ -228,7 +231,7 @@ class EmptyDirTest extends Unit
      */
     public function should_throw_if_scaffolding_fails_due_to_file_copy(): void
     {
-        $this->uopzSetStaticMethodReturn(FS::class, 'recurseCopy', false);
+        $this->setMethodReturn(FS::class, 'recurseCopy', false);
 
         $wpRootDir = FS::tmpDir('empty-dir_');