+++ tests/unit/lucatume/WPBrowser/WordPress/WpConfigFileGeneratorTest.php
@@ -4,13 +4,16 @@ namespace lucatume\WPBrowser\WordPress;
 
 use Codeception\Test\Unit;
 use lucatume\WPBrowser\Tests\Traits\TmpFilesCleanup;
-use lucatume\WPBrowser\Tests\Traits\UopzFunctions;
+use lucatume\WPBrowser\Traits\UopzFunctions;
 use lucatume\WPBrowser\Utils\Env;
 use lucatume\WPBrowser\Utils\Filesystem as FS;
 use lucatume\WPBrowser\WordPress\Database\MysqlDatabase;
 use lucatume\WPBrowser\WordPress\Database\SQLiteDatabase;
 use tad\Codeception\SnapshotAssertions\SnapshotAssertions;
 
+/**
+ * @group slow
+ */
 class WpConfigFileGeneratorTest extends Unit
 {
     use UopzFunctions;
@@ -58,7 +61,7 @@ class WpConfigFileGeneratorTest extends Unit
             'wp-settings.php' => '<?php echo "Hello there!";',
             'wp-config-sample.php' => '<?php echo "Hello there!";'
         ]);
-        $this->uopzSetFunctionReturn('file_get_contents', static function ($file) use ($wpRootDir) {
+        $this->setFunctionReturn('file_get_contents', static function ($file) use ($wpRootDir) {
             return $file === $wpRootDir . '/wp-config-sample.php' ? false : file_get_contents($file);
         }, true);